refactor: imageInfo 원시값 포장
