--- src/main/java/nextstep/courses/domain/SessionStatus.java
@@ -1,18 +1,0 @@
-package nextstep.courses.domain;
-
-public enum SessionStatus {
-	PREPARING(0),
-	PROCEEDING(1),
-	CLOSED(2);
-
-	private int statusValue;
-
-	SessionStatus(int statusValue) {
-		this.statusValue = statusValue;
-	}
-
-	public int getStatusValue() {
-		return statusValue;
-	}
-
-}
